feat: calculator 기능 추가

+++ src/main/kotlin/stringCalculator/Operator.kt
@@ -6,4 +6,4 @@ fun substrate(a: Int, b: Int): Int = a - b
 
 fun multiple(a: Int, b: Int): Int = a * b
 
-fun divide(a: Int, b: Int): Double = a.toDouble() / b
+fun divide(a: Int, b: Int): Int = a / b
